patra_agent/cloud_services/model-placement-agent/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40

# LLM providers
anthropic>=0.18.0
openai>=1.0.0
ollama>=0.1.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0
httpx>=0.25.0
